Renamed "js/index.ts" to "ts/index.ts"
