+++ wasm/bootstrap.js
@@ -1,2 +1,2 @@
 // We need a seam so the WebAssembly can be imported asynchronously
-import("./js/index.ts").catch(console.error);
+import("./ts/index.ts").catch(console.error);
