+++ wasm/ts/index.ts
